Move HTML5toTouch file location and Remove MultiBackend
